+++ lib/settings.d.ts
@@ -1,6 +1,6 @@
 import z from '@deepseek-ai/schemastery';
 export * from './shared/settings.js';
-export declare const ASUKA_SETTINGS_NAMESPACE: Branded<"SettingsNamespace">;
+export declare const ASUKA_SETTINGS_NAMESPACE: import("@deepseek-ai/dsh-settings").SettingsNamespace;
 export declare const AsukaThemeSettingsSchema: z<Schemastery.ObjectS<{
     mode: z<"off" | "after-class" | "tokyo3-night", "off" | "after-class" | "tokyo3-night">;
     wallpaperEnabled: z<boolean, boolean>;